Add files via upload
A message system that can be used by a single user connecting to the server at the same time with Java. Users connect to the server at different times and send messages to each other. When the new user logs in, they can check if they have a message, send a new message or end the session.